+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Full changelog coming soon. For now, here are the most important changes for doing major migrations:
 
+## Migrating to version 1.1
+- `OpenSeaPort::computeFees()` No longer accepts a `fees` parameter, relying solely on `asset`.
+
 ## Migrating to version 1.0
 
 Version 1.0 introduces bundling for semi-fungible and fungible assets, serialized asset collections, a smaller bundle size, and more, along with many bug fixes.
@@ -12,11 +15,13 @@ Version 1.0 introduces bundling for semi-fungible and fungible assets, serialize
 
 **Breaking type changes**
 - `profile_img_url` in the `OpenSeaAccount` type has been renamed to `profileImgUrl`
+- `OpenSeaPort::computeFees()` now returns a new type, `ComputedFees`, which expands "BPS" to "BasisPoints" for each of its member keys
 
 **Breaking interface changes**
-- `OpenSeaPort::computeFees()` now takes in either the `fees` from a collection or a single asset, and returns a new and more consistent type, `ComputedFees`
+- `OpenSeaPort::computeFees()` now takes in a single asset, and returns a new and more consistent type, `ComputedFees`
 - `schemaName` has been moved out of main method calls and into the `Asset` type
 - `OpenSeaAPI::getAsset()` now accepts named arguments, consistent with other methods
+- `OpenSeaPort::createFactorySellOrders` now accepts an `assets` parameter consistent with the other `Asset` type parameters in other methods
 - `OpenSeaPort::getTokenBalance()` no longer defaults to the WETH address if no `tokenAddress` is set. And it accepts a `schemaName` parameter instead of an ABI
 - `OpenSeaPort::approveNonFungibleToken()` has been renamed to `approveSemiOrNonFungibleToken()`, though it always worked for semi-fungible tokens like ERC-1155